Remove Metadata class and substitute for dict.
Considering we only ever need one instance of the metadata information,
it is silly to require the full overhead of a class. Additionally,
classes can make it difficult to append attributes across function calls
which could be necessary in the future.

A dict was chosen instead as it is lightweight, has named attributes for
readability (like a class), and is more easily mutatable.